Game Master App/requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
colorama==0.4.6
rich==13.7.0
typer==0.9.0 